Update: read logs, get topic, get device

+++ Dashboard_Web/models/iot/device.py
@@ -34,12 +34,18 @@ class Device(db.Model):
         
         return create_with_integrity(new_device, Device.__tablename__)
     
+    def get_single_device(id):
+        device = Device.query.filter(Device.device_id == id).first()
+        if device is not None:
+            return device    
+    
     def get_devices_with_topics():
         devices = Device.query.options(joinedload(Device.topic)).all()
         return devices
     
     def delete_device(device_id):
-        device = Device.query.get(Device.id == device_id).first()
-        db.session.delete(device)
-        db.session.commit()
-        return Device.get_sensors()
+        device = Device.query.get(device_id)
+        if device:
+            db.session.delete(device)
+            db.session.commit()
+        return Device.get_devices_with_topics()
